server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
psycopg==3.2.2
psycopg-binary==3.2.2
python-dotenv==1.0.1
openai>=1.0.0
pydantic==2.9.0
httpx>=0.27.0
requests>=2.31.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
